requirements.txt
python-telegram-bot==20.5
aiohttp==3.9.3
torch==2.2.1
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
fastapi==0.110.2
uvicorn==0.29.0
diffusers==0.27.2
transformers==4.40.1
realesrgan==0.2.5
